11_packages/examples/0_slide-examples/requirements.txt
python-dotenv==1.2.1
yattag==1.16.1
